Adds darwin/os x specific process.spawn code
After trying the ln -s "trick" that was listed in the vlc-client readme, it broke my VLC installation on my mac.
deleting the created symlinks made it work again.
Therefore I added a clause to allow macs to also spawn processes from within the mac. This is a small fix for my specific project.